:art: Cleanup up unused classes and moving code inside right packages
