+++ src/main/java/org/terracotta/statistics/StatisticType.java
@@ -13,31 +13,17 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.terracotta.statistics.extended;
+package org.terracotta.statistics;
 
 /**
- * @author Ludovic Orban
+ * @author Mathieu Carbou
  */
-public interface Result {
+public enum StatisticType {
 
-  /**
-   * Count.
-   *
-   * @return the statistic
-   */
-  SampledStatistic<Long> count();
+  COUNTER,
+  RATE,
+  RATIO,
+  GAUGE,
+  TABLE
 
-  /**
-   * Rate.
-   *
-   * @return the statistic
-   */
-  SampledStatistic<Double> rate();
-
-  /**
-   * Latency.
-   *
-   * @return the latency
-   */
-  Latency latency();
 }
