+++ src/main/java/org/terracotta/statistics/TableValueStatistic.java
@@ -15,8 +15,6 @@
  */
 package org.terracotta.statistics;
 
-import org.terracotta.statistics.extended.StatisticType;
-
 import java.io.Serializable;
 import java.util.Collection;
 import java.util.function.BiConsumer;
