+++ src/main/java/org/terracotta/statistics/derived/EventParameterSimpleMovingAverage.java
@@ -27,10 +27,10 @@ import java.util.concurrent.atomic.LongAccumulator;
 import java.util.concurrent.atomic.LongAdder;
 
 import org.terracotta.statistics.ValueStatistic;
-import org.terracotta.statistics.extended.StatisticType;
+import org.terracotta.statistics.StatisticType;
 import org.terracotta.statistics.observer.ChainedEventObserver;
 
-import static org.terracotta.statistics.extended.StatisticType.GAUGE;
+import static org.terracotta.statistics.StatisticType.GAUGE;
 
 /**
  *
