--- src/main/java/org/terracotta/statistics/extended/Latency.java
@@ -1,43 +1,0 @@
-/*
- * Copyright Terracotta, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.terracotta.statistics.extended;
-
-/**
- * @author Ludovic Orban
- */
-public interface Latency {
-
-  /**
-   * Minimum latency observed.
-   *
-   * @return Minimum observed latency. NULL if no operation was observed.
-   */
-  SampledStatistic<Long> minimum();
-
-  /**
-   * Maximum latency observed.
-   *
-   * @return Maximum observed latency. NULL if no operation was observed.
-   */
-  SampledStatistic<Long> maximum();
-
-  /**
-   * Average observed latency.
-   *
-   * @return Average observed latency. NULL if no operation was observed.
-   */
-  SampledStatistic<Double> average();
-}
