+++ src/main/java/org/terracotta/statistics/registry/ValueStatisticDescriptor.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.terracotta.context.extended;
+package org.terracotta.statistics.registry;
 
 import java.util.Arrays;
 import java.util.Collections;
